+++ inst/www/react-tools/react-tools.js
@@ -157,7 +157,9 @@ function defaultReceiveMessage(el, _ref) {
 }
 
 var defaultOptions = {
-  receiveMessage: defaultReceiveMessage
+  receiveMessage: defaultReceiveMessage,
+  type: false,
+  ratePolicy: null
 };
 /**
  * Installs a new Shiny input binding based on a React component.
@@ -172,6 +174,28 @@ var defaultOptions = {
  * - receiveMessage: Implementation of Shiny.InputBinding to use in place of
  *   the default. Typically overridden as an optimization to perform
  *   incremental value updates.
+ * - type: `false`, a string, or a function.
+ *     - `false` (the default): denotes that the value produced by this input
+ *       should not be intercepted by any handlers registered in R on the
+ *       server using shiny::registerInputHandler().
+ *     - string: denotes the input's *type* and should correspond to the
+ *       type parameter of shiny::registerInputHandler().
+ *     - function: A function called with `this` bound to the InputBinding
+ *       instance and passed a single argument, the input's containing DOM
+ *       element. The function should return either `false` or a string
+ *       corresponding to the type parameter of shiny::registerInputHandler().
+ * - ratePolicy: A rate policy object as defined in the documentation for
+ *     getRatePolicy(): https://shiny.rstudio.com/articles/building-inputs.html
+ *     A rate policy object has two members:
+ *     - `policy`: Valid values are the strings "direct", "debounce", and
+ *       "throttle". "direct" means that all events are sent immediately.
+ *     - `delay`: Number indicating the number of milliseconds that should be
+ *       used when debouncing or throttling. Has no effect if the policy is
+ *       direct.
+ *     The specified rate policy is only applied when `true` is passed as the
+ *     second argument to the `setValue` function passed as a prop to the
+ *     input component.
+ *
  */
 
 function reactShinyInput(selector, name, component, options) {
@@ -204,8 +228,9 @@ function reactShinyInput(selector, name, component, options) {
     }, {
       key: "setValue",
       value: function setValue(el, value) {
+        var rateLimited = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : false;
         this.setInputValue(el, value);
-        this.getCallback(el)();
+        this.getCallback(el)(rateLimited);
         this.render(el);
       }
     }, {
@@ -223,7 +248,6 @@ function reactShinyInput(selector, name, component, options) {
     }, {
       key: "unsubscribe",
       value: function unsubscribe(el, callback) {
-        jquery__WEBPACK_IMPORTED_MODULE_3___default()(el).removeData('callback');
         react_dom__WEBPACK_IMPORTED_MODULE_1___default.a.render(null, el);
       }
     }, {
@@ -231,10 +255,26 @@ function reactShinyInput(selector, name, component, options) {
       value: function receiveMessage(el, data) {
         options.receiveMessage.call(this, el, data);
       }
+    }, {
+      key: "getType",
+      value: function getType(el) {
+        if (typeof options.type === 'function') {
+          return options.type.call(this, el);
+        } else if (options.type === false || typeof options.type === 'string') {
+          return options.type;
+        } else {
+          throw new Error('options.type must be false, a string, or a function');
+        }
+      }
+    }, {
+      key: "getRatePolicy",
+      value: function getRatePolicy() {
+        return options.ratePolicy;
+      }
       /*
        * Methods not present in Shiny.InputBinding but accessible to users
        * through `this` in receiveMessage
-       * */
+       */
 
     }, {
       key: "getInputValue",