+++ NEWS.md
@@ -1,3 +1,13 @@
+# reactR 0.4.1
+
+* Add support for `shiny::registerInputHandler` in Shiny inputs; [pull 28](https://github.com/react-R/reactR/pull/28)
+
+* Add support for Shiny rate limit [pull 29](https://github.com/react-R/reactR/pull/29)
+
+* Update react to `16.8.6`
+
+* Add `mobx` dependencies available through `html_dependency_mobx()`
+
 # reactR 0.4.0
 
 * Add Shiny input scaffold and functionality; [tutorial](https://react-r.github.io/reactR/articles/intro_inputs.html) and [pull 22](https://github.com/react-R/reactR/pull/22) thanks @alandipert